Refactor Control.tsx file and layout fixes for GroupControl modal (#552)
* Layout fixes

* Move ControlItemSelect in separate file

* Fix getDistincts

* Fix styles

* Move ControlItemSelect to renderSelectControl function
