+++ src/ui/units/dash/containers/Dialogs/GroupControl/GroupControl.scss
@@ -12,8 +12,9 @@
         display: flex;
         flex-direction: column;
         height: 100%;
-        width: 255px;
+        width: 262px;
         max-height: 100%;
+        padding-top: 2px;
     }
 
     &__selectors-list {
